Code files/Back-end/GCP Spend-Bot/requirements.txt
sqlglot>=23.0.0
google-cloud-bigquery
google-genai
